Extracted idb logic from GlobalStore into IdbAdapter for testability
