+++ source/frontend/app/adapters/IdbAdatper.test.js
@@ -1,0 +1,9 @@
+jest.mock('idb');
+
+import idb from 'idb';
+
+import * as idbAdapter from './IdbAdapter';
+
+it('idbAdapter should be defined', () => {
+  expect(idbAdapter).toBeDefined();
+});
