+++ source/frontend/app/stores/GlobalStore.js
@@ -1,14 +1,13 @@
 import { EventEmitter } from 'events';
 import dispatcher from '../dispatcher';
-import idb from 'idb';
 
 import * as api from '../adapters/ZhawoAdapter';
+import * as idbAdapter from '../adapters/IdbAdapter';
 
 class GlobalStore extends EventEmitter {
   constructor() {
     super();
     this.theme = 'lightTheme';
-
     this.currentUser = '';
     this.currentUserType = '';
     this.possibleNames = [];
@@ -22,15 +21,13 @@ class GlobalStore extends EventEmitter {
   async handleActions(action) {
     switch (action.type) {
       case 'SET_CURRENT_USER':
-        this.currentUser = action.payload.name;
-        this.currentUserType = action.payload.type;
         this.setCurrentUser(action.payload.name, action.payload.type);
-        this.emit('current_user_changed');
         break;
       case 'TOGGLE_DRAWER':
         this.drawerOpen = !this.drawerOpen;
         this.emit('drawerOpen_changed');
         break;
+
       case 'GET_POSSIBLE_NAMES':
         const possibleNames = await api.getPossibleNames().catch(err => {
           console.error(err);
@@ -75,93 +72,34 @@ class GlobalStore extends EventEmitter {
     } else {
       this.theme = 'lightTheme';
     }
-
     this.setThemeInDB(this.theme);
   }
 
-  async getDBInstance() {
-    return new Promise(async (resolve, reject) => {
-      let dbInstance = await idb.open('zhawoDB', 1, function(upgradeDB) {
-        switch (upgradeDB.oldVersion) {
-          case 0:
-            upgradeDB.createObjectStore('info', { keyPath: 'id' });
-          case 1:
-          // When we make a version 2 we can add those features here
-        }
-      });
-      if (dbInstance) resolve(dbInstance);
-    });
-  }
-
   async getUsernameFromDB() {
-    let dbInstance = await this.getDBInstance();
-
-    let tx = dbInstance.transaction('info', 'readonly');
-    let store = tx.objectStore('info');
-
-    // add, clear, count, delete, get, getAll, getAllKeys, getKey, put
-    let user = await store.get('username');
-
+    let user = await idbAdapter.getUsername();
     this.currentUser = user.username;
     this.currentUserType = user.type;
-
     this.emit('current_user_changed');
-    dbInstance.close();
   }
 
   async getThemeFromDB() {
-    let dbInstance = await this.getDBInstance();
-
-    let tx = dbInstance.transaction('info', 'readonly');
-    let store = tx.objectStore('info');
-
-    // add, clear, count, delete, get, getAll, getAllKeys, getKey, put
-    let theme = await store.get('theme');
-
+    let theme = await idbAdapter.getTheme();
     if (theme) this.theme = theme.theme;
     this.emit('theme_changed');
-
-    dbInstance.close();
   }
 
-  // TODO: change so that what you are saving is the key
-
   async setThemeInDB(theme) {
-    let dbInstance = await this.getDBInstance();
-
-    let tx = dbInstance.transaction('info', 'readwrite');
-    let store = tx.objectStore('info');
-
-    await store.put({ id: 'theme', theme: theme });
-
-    await tx.complete;
-    dbInstance.close();
+    await idbAdapter.setTheme(theme);
   }
 
   async setCurrentUser(name, type) {
-    let dbInstance = await this.getDBInstance();
-
-    let tx = dbInstance.transaction('info', 'readwrite');
-    let store = tx.objectStore('info');
-
-    await store.put({ id: 'username', username: name, type: type });
-
-    await tx.complete;
-    dbInstance.close();
+    this.currentUser = name;
+    this.currentUserType = type;
+    await idbAdapter.setUser(name, type);
   }
 
   async removeCurrentUser() {
-    let dbInstance = await this.getDBInstance();
-
-    let tx = dbInstance.transaction('info', 'readwrite');
-    let store = tx.objectStore('info');
-
-    let allSavedItems = await store.getAllKeys();
-
-    await store.delete('username');
-
-    await tx.complete;
-    dbInstance.close();
+    await idbAdapter.removeUser();
   }
 }
